Refactor : `Repository` class 이름 수정
