+++ src/main/java/kr/codesqaud/cafe/repository/ArticleRepository.java
@@ -1,28 +1,14 @@
 package kr.codesqaud.cafe.repository;
 
 import kr.codesqaud.cafe.domain.Article;
-import org.springframework.stereotype.Repository;
 
-import java.util.*;
-import java.util.concurrent.ConcurrentLinkedDeque;
-import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.stream.Collectors;
+import java.util.List;
+import java.util.Optional;
 
-@Repository
-public class ArticleRepository {
-    private final Queue<Article> store = new ConcurrentLinkedQueue<>();
+public interface ArticleRepository {
+    void save(Article article);
 
-    public void save(Article article) {
-        store.add(article);
-    }
+    Optional<Article> findById(Long id);
 
-    public Optional<Article> findById(Long id) {
-        return store.stream()
-                .filter(e -> e.equals(id))
-                .findAny();
-    }
-
-    public List<Article> findAll() {
-        return List.copyOf(store);
-    }
+    List<Article> findAll();
 }
